+++ src/Evolution/Systems/GeneralizedHarmonic/UpwindPenaltyCorrection.cpp
@@ -8,6 +8,7 @@
 
 #include "DataStructures/DataBox/PrefixHelpers.hpp"
 #include "DataStructures/DataVector.hpp"
+#include "DataStructures/Tags/TempTensor.hpp"
 #include "DataStructures/Tensor/Tensor.hpp"
 #include "DataStructures/Variables.hpp"
 #include "Utilities/GenerateInstantiations.hpp"
